requirements.txt
pandas~=2.2.3
openpyxl~=3.1.5
python-dotenv~=1.0.1
openai~=1.59.6
streamlit~=1.41.1
pillow~=11.1.0
spacy~=3.8.4
scikit-learn~=1.6.1
numpy~=2.2.1
langdetect~=1.0.9
camelot-py~=1.0.0
pdfplumber~=0.11.5